// packages/conciliaciones/src/enums.d.ts
export declare enum RunStatus {
    OPEN = "OPEN",
    CLOSED = "CLOSED"
}
export declare enum RunMemberRole {
    OWNER = "OWNER",
    EDITOR = "EDITOR",
    VIEWER = "VIEWER"
}
export declare enum UnmatchedSystemStatus {
    OVERDUE = "OVERDUE",
    DEFERRED = "DEFERRED"
}
export declare enum PendingStatus {
    OPEN = "OPEN",
    IN_PROGRESS = "IN_PROGRESS",
    RESOLVED = "RESOLVED"
}
export declare enum ChequeStatus {
    ISSUED = "ISSUED",
    CLEARED = "CLEARED",
    OVERDUE = "OVERDUE"
}
//# sourceMappingURL=enums.d.ts.map