// packages/conciliaciones/src/entities/unmatched-extract.entity.d.ts
import { ReconciliationRunEntity } from './reconciliation-run.entity';
import { ExtractLineEntity } from './extract-line.entity';
export declare class UnmatchedExtractEntity {
    id: string;
    runId: string;
    run: ReconciliationRunEntity;
    extractLineId: string;
    extractLine: ExtractLineEntity;
}
//# sourceMappingURL=unmatched-extract.entity.d.ts.map